airflow/requirements.txt
requests
psycopg2-binary
SQLAlchemy
